refactor: Move new utility functions and associated tests into ui-react-rx/src/util
