+++ packages/ui-react-rx/src/util/isTestChain.spec.js
@@ -1,8 +1,8 @@
-// Copyright 2017-2018 @polkadot/apps authors & contributors
+// Copyright 2017-2018 @polkadot/ui-react-rx authors & contributors
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import { isTestChain } from './NodeInfo';
+import isTestChain from './isTestChain';
 
 describe('check chain spec to configure keyring test mode and dev accounts availability', () => {
   it('enables test environment when chain specification matches text of dev or loc(al)', () => {
